Terminadas views

+++ app/views.py
@@ -5,12 +5,9 @@ from . import models
 import re
 import logging
 from django.contrib.auth.decorators import login_required
+from .models import Person
 
 logger = logging.getLogger(__name__)
-
-
-# Create your views here.
-from .models import Person
 
 
 def home(request):
@@ -25,42 +22,51 @@ def home(request):
         persons = ""
 
     objects = {
-        "persons": persons
+        "persons": persons,
+        "nav_search": person_name_query
     }
+
     return render(request, "index.html", objects)
 
 
-#@login_required
+@login_required
 def add(request):
     form = models.PersonForm()
+    user = request.user.email
+    admin = "admin@admin.es"
 
     if request.method == "POST":
         form = models.PersonForm(request.POST, request.FILES)
 
         if form.is_valid():
-            firstname = form.cleaned_data["firstname"]
-            lastname = form.cleaned_data["lastname"]
-            email = form.cleaned_data["email"]
-            image = request.FILES
-            person = Person(firstname=firstname, lastname=lastname, email=email)
-            person.save()
+            if user == admin:
+                firstname = form.cleaned_data["firstname"]
+                lastname = form.cleaned_data["lastname"]
+                email = form.cleaned_data["email"]
+                image = request.FILES
+                person = Person(firstname=firstname, lastname=lastname, email=email)
+                person.save()
 
-            if image:
-                image_name = "app/static/data/img/{}.jpg".format(person.id)
-                fs = FileSystemStorage()
-                fs.save(image_name, image["image"])
+                if image:
+                    image_name_2_save = "app/static/data/img/{}.jpg".format(person.id)
+                    image_name = "data/img/{}".format(person.id)
+                    fs = FileSystemStorage()
+                    fs.save(image_name_2_save, image["image"])
+                else:
+                    image_name = Person.objects[0].image
+
+                person.image = image_name
+                person.save()
+                logger.info("New person added")
+
+                return redirect("/")
             else:
-                image_name = Person.objects[0].image
-
-            person.image = image_name
-            person.save()
-            logger.info("New person added")
-
-            return redirect("/")
+                return redirect("/")
 
     objects = {
         "form": form
     }
+
     return render(request, "add.html", objects)
 
 
@@ -70,12 +76,16 @@ def detail(request, id):
     objects = {
         "person": person
     }
+
     return render(request, "detail.html", objects)
 
 
 @login_required
 def edit(request, id):
     person = Person.objects.get(id=id)
+
+    user = request.user.email
+    admin = "admin@admin.es"
 
     if request.method == "GET":
         form = models.PersonForm()
@@ -84,15 +94,19 @@ def edit(request, id):
         form = models.PersonForm(data=request.POST)
 
         if form.is_valid():
-            firstname = form.cleaned_data["firstname"]
-            lastname = form.cleaned_data["lastname"]
-            email = form.cleaned_data["email"]
+            if user == admin:
+                firstname = form.cleaned_data["firstname"]
+                lastname = form.cleaned_data["lastname"]
+                email = form.cleaned_data["email"]
 
-            person.firstname = firstname
-            person.lastname = lastname
-            person.email = email
-            person.save()
-            logger.info("The person was edited")
+                person.firstname = firstname
+                person.lastname = lastname
+                person.email = email
+                person.save()
+                logger.info("The person was edited")
+
+            else:
+                return redirect("/")
 
         else:
             logger.error("Form not valid")
@@ -103,19 +117,28 @@ def edit(request, id):
         "form": form,
         "person": person
     }
+
     return render(request, "edit.html", objects)
 
 
-#@login_required
+@login_required
 def delete(request, id):
-    person = Person.objects(id=id)
-    person.delete()
-    logger.info("The person was deleted")
+    user = request.user.email
+    admin = "admin@admin.es"
 
-    objects = {
-        "delete": id
-    }
-    return JsonResponse(objects)
+    if user == admin:
+        person = Person.objects(id=id)
+        person.delete()
+        logger.info("The person was deleted")
+
+        objects = {
+            "delete": id
+        }
+
+        return JsonResponse(objects)
+
+    else:
+        return redirect("/")
 
 
 def login(request):
